+++ blitzcc/exprnode.cpp
@@ -75,14 +75,14 @@ TNode *CastNode::translate( Codegen *g ){
 // Sequence of Expressions //
 /////////////////////////////
 void ExprSeqNode::semant( Environ *e ){
-	for( int k=0;k<exprs.size();++k ){
+	for( int k=0;k<(int)exprs.size();++k ){
 		if( exprs[k] ) exprs[k]=exprs[k]->semant( e );
 	}
 }
 
 TNode *ExprSeqNode::translate( Codegen *g,bool userlib ){
 	TNode *t=0,*l=0;
-	for( int k=0;k<exprs.size();++k ){
+	for( int k=0;k<(int)exprs.size();++k ){
 
 		TNode *q=exprs[k]->translate(g);
 
@@ -111,10 +111,10 @@ TNode *ExprSeqNode::translate( Codegen *g,bool userlib ){
 }
 
 void ExprSeqNode::castTo( DeclSeq *decls,Environ *e,bool userlib ){
-	if( exprs.size()>decls->size() ) ex( "Too many parameters" );
+	if( (int)exprs.size()>decls->size() ) ex( "Too many parameters" );
 	for( int k=0;k<decls->size();++k ){
 		Decl *d=decls->decls[k];
-		if( k<exprs.size() && exprs[k] ){
+		if( k<(int)exprs.size() && exprs[k] ){
 
 			//USERLIB!
 			if( userlib && d->type->structType() ){
@@ -132,14 +132,14 @@ void ExprSeqNode::castTo( DeclSeq *decls,Environ *e,bool userlib ){
 		}else{
 			if( !d->defType ) ex( "Not enough parameters" );
 			ExprNode *expr=constValue( d->defType );
-			if( k<exprs.size() ) exprs[k]=expr;
+			if( k<(int)exprs.size() ) exprs[k]=expr;
 			else exprs.push_back( expr );
 		}
 	}
 }
 
 void ExprSeqNode::castTo( Type *t,Environ *e ){
-	for( int k=0;k<exprs.size();++k ){
+	for( int k=0;k<(int)exprs.size();++k ){
 		exprs[k]=exprs[k]->castTo( t,e );
 	}
 }
@@ -220,7 +220,7 @@ int IntConstNode::intValue(){
 }
 
 float IntConstNode::floatValue(){
-	return value;
+	return (float)value;
 }
 
 string IntConstNode::stringValue(){
@@ -308,7 +308,7 @@ ExprNode *UniExprNode::semant( Environ *e ){
 			case '+':e=d_new FloatConstNode( +c->floatValue() );break;
 			case '-':e=d_new FloatConstNode( -c->floatValue() );break;
 			case ABS:e=d_new FloatConstNode( c->floatValue()>=0 ? c->floatValue() : -c->floatValue() );break;
-			case SGN:e=d_new FloatConstNode( c->floatValue()>0 ? 1 : (c->floatValue()<0 ? -1 : 0) );break;
+			case SGN:e=d_new FloatConstNode( (float)(c->floatValue()>0 ? 1 : (c->floatValue()<0 ? -1 : 0)) );break;
 			}
 		}
 		delete this;
