+++ blitzcc/stmtnode.cpp
@@ -49,7 +49,7 @@ void StmtSeqNode::reset( const string &file,const string &lab ){
 // Statement Sequence //
 ////////////////////////
 void StmtSeqNode::semant( Environ *e ){
-	for( int k=0;k<stmts.size();++k ){
+	for( int k=0;k<(int)stmts.size();++k ){
 		try{ stmts[k]->semant( e ); }
 		catch( Ex &x ){ 
 			if( x.pos<0 ) x.pos=stmts[k]->pos;
@@ -63,7 +63,7 @@ void StmtSeqNode::translate( Codegen *g ){
 
 	string t=fileLabel;
 	fileLabel=file.size() ? fileMap[file] : "";
-	for( int k=0;k<stmts.size();++k ){
+	for( int k=0;k<(int)stmts.size();++k ){
 		StmtNode *stmt=stmts[k];
 		stmt->debug( stmts[k]->pos,g );
 		try{
@@ -552,7 +552,7 @@ void SelectNode::semant( Environ *e ){
 	Decl *d=e->decls->insertDecl( genLabel(),expr->sem_type,DECL_LOCAL );
 	sem_temp=d_new DeclVarNode( d );
 
-	for( int k=0;k<cases.size();++k ){
+	for( int k=0;k<(int)cases.size();++k ){
 		CaseNode *c=cases[k];
 		c->exprs->semant( e );
 		c->exprs->castTo( ty,e );
@@ -572,7 +572,7 @@ void SelectNode::translate( Codegen *g ){
 	vector<string> labs;
 	string brk=genLabel();
 
-	for( int k=0;k<cases.size();++k ){
+	for( int k=0;k<(int)cases.size();++k ){
 		CaseNode *c=cases[k];
 		labs.push_back( genLabel() );
 		for( int j=0;j<c->exprs->size();++j ){
@@ -585,7 +585,7 @@ void SelectNode::translate( Codegen *g ){
 	}
 	if( defStmts ) defStmts->translate( g );
 	g->code( jump( brk ) );
-	for( int k=0;k<cases.size();++k ){
+	for( int k=0;k<(int)cases.size();++k ){
 		CaseNode *c=cases[k];
 		g->label( labs[k] );
 		c->stmts->translate( g );
